% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lasvmPredictWrapper}
\alias{lasvmPredictWrapper}
\title{lasvmPredictWrapper}
\usage{
lasvmPredictWrapper(x, SV, elif, gamma, kdegree, kcoef0, bias, kerneltype,
  verbose = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{SV}{matrix of support vectors}

\item{elif}{vector of alphas}

\item{gamma}{gamma of RBF kernel}

\item{kdegree}{degree for POLY kernel}

\item{kcoef0}{coef0 for kernel}

\item{bias}{bias term}

\item{kerneltype}{type of kernel to use}

\item{verbose}{verbose output?}
}
\value{
a list consisting of
	predictions		just the predictions
}
\description{
Use lasvm to predict a given problem.
}

