% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simanle.R
\docType{package}
\name{simanle}
\alias{simanle}
\alias{simanle-package}
\title{simanle: A Simple Wrapper for the LASVM Solver}
\description{
This is a simple wrapper for LASVM Solver (see http://leon.bottou.org/projects/lasvm).
LASVM is an online variant of the SMO solver.
}
\section{simanle functions}{

lasvmTrain
lasvmPredict
}

