% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lasvmTrainWrapper}
\alias{lasvmTrainWrapper}
\title{lasvmTrainWrapper}
\usage{
lasvmTrainWrapper(x, y, gamma, cost, degree = 3, coef0 = 0L,
  optimizer = 1L, kernel = 2L, selection = 0L, termination = 0L,
  sample = 0, cachesize = 256L, bias = 1L, epochs = 1L,
  epsilon = 0.001, verbose = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{y}{training labels}

\item{gamma}{RBF kernel bandwidth}

\item{cost}{regularization constant}

\item{degree}{degree for poly kernel}

\item{coef0}{coefficient for poly kernel}

\item{optimizer}{type of optimizer}

\item{kernel}{kernel type}

\item{selection}{selection strategy}

\item{termination}{criterion for stopping}

\item{sample}{parameter for stopping criterion, e.g. seconds}

\item{cachesize}{size of kernel cache}

\item{bias}{use  bias?}

\item{epochs}{number of epochs}

\item{epsilon}{stopping criterion parameter}

\item{verbose}{verbose output?}
}
\value{
a list consisting of
	SV		matrix of support vectors
	alpha		vector of alpha coefficients
	bias		bias term
}
\description{
Use lasvm to train a given problem.
}

